Bumper fp-kontrakter til 6.2.1: package rename endring for fpwsproxy avhengigheter
